requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
celery
redis
python-dotenv
requests
slowapi
python-multipart
email-validator
gunicorn
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
pytest-cov
httpx
requests-mock